+++ Pester/Individualfunction.tests.ps1
@@ -1,0 +1,7 @@
+
+$individualfunctionslocation = Join-Path -Path (Join-Path ".././" -ChildPath "PSCloudPC") -ChildPath "Src/Tests"
+
+foreach ($file in Get-ChildItem $individualfunctionslocation) {
+
+    Invoke-Pester -Path $file.FullName -Output Detailed
+}