+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 This file contains all the changes to this Windows 365 PowerShell module.
+
+## 1.0.15
+###Changes
+- Restructure module repository
+- Changed Authentication to Microsoft.Graph
+- Added authentication for token based auth
+- Added Pester tests for disconnect-windows365
+
+## 1.0.13 [25-09-2024]
+###Changes
+- Changed authentication check the docs for examples
+- Added Certs based authentication
+- Added Paging for Graph Requests
+
 ## 1.0.4 [20-1-2023]
 ### Bugfixes
 - Bug fixed in Update-CPCUserSettingsPolicy, parameter value was not working